@@ -2,7 +2,8 @@ import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 
-const url = 'http://localhost:3000/cousers'
+const url = 'http://localhost:3000/cousers';
+const url2 = ' http://localhost:3000/Author';
 @Injectable({
   providedIn: 'root'
 })
@@ -16,14 +17,17 @@ export class HomeService {
     return this.http.get<any[]>(`${url}?status_2=true`)
   }
   getCourseHome_Instructor(): Observable<any[]> {
-    return this.http.get<any[]>(`${url}?status_3=true`)
+    return this.http.get<any[]>(`${url2}?status_3=true`)
   }
   getCourseHome_Student(): Observable<any[]> {
-    return this.http.get<any[]>(`${url}?status_4=true`)
+    return this.http.get<any[]>(`${url2}?status_4=true`)
   }
+<<<<<<< HEAD
 
 
 
 
 
+=======
+>>>>>>> c4ed848db735d581035374aec130dc5a24fdc658
 }
